Fix #926
`utils::tokenize_quoted()` is supposed to un-escape symbols inside
double quotes. For example, the code
`utils::tokenize_quoted("\"\x\\\"")` should return a string "x\" (letter
x and a single backslash). As this example demonstrates, to get
a literal backslash, one has to write two backslashes in a row.

The function is structured such that it first founds a portion of
a string that's enclosed in the double quotes, and then processes it
character-by-character. At every character, it takes a look at the previous
one to see if it's a backslash, and if it is, it prints out a character
corresponding to the escaped sequence (e.g. "\n" turns into a newline).

This doesn't work for long runs of backslashes, though: if the input is
"\\\\" (four backslashes), the function would:
1. take a look at the first backslash and ignore it since there is
   nothing before it;
2. take a look at the second backslash, then the first one, realize it's
   an escape sequence, and add one backslash to the output;
3. take a look at the third backslash, then the second one, **wrongly**
   conclude that it's an escape sequence, and add one backslash to the
   output;
4. take a look at the 4th, then 3rd backslash, correctly conclude that
   this is an escape sequence, and print out one backslash to the
   output.

The result is three backslashes instead of two.

To combat that, original code kept a boolean, which flipped every time
when a backslash was printed. So on step 2 we printed something and
flipped the boolean, then on step three we didn't print anything because
the boolean was wrong (but we flipped it again), and then on step 4 we
printed a backslash again.

This logic worked fine on consecutive runs of backslashes, but broke
when escaped backslash was followed by something else, then escaped
backslash again. The boolean was in the wrong state, and the second
escaped backslash was simply ignored.

In order to fix this, I changed the function to immediately check the
following character once it encountered a backslash, *and* take a double
step afterwards. This means that a string like "\\x\\" is processed in
three iterations rather than 5:

1. check first char (backslash), look on the second char (backslash),
   output backslash, advance position by two characters;
2. check third char (x), output "x", advance position by one char;
3. check fourth char (backslash), check fifth char (backslash), output
   backslash, advance position by two chars.