FeedListFormAction: Remove non-functional resize detection code
The width was retrieved from stfl with variable name "items:w".
There is no element named "items" in the `feedlist.stfl` file so `width`
would always be equal to 0, resulting in no action on an actual resize.

It is probably best to fix this in a generic way when implementing [0].

[0] https://github.com/newsboat/newsboat/issues/390