Move some tests from Matcher to FilterParser
We don't examine the parse tree in C++, because that parser is going
away soon. The only thing we care about is whether it can parse a given
input or not. To ensure that it's parsed correctly, we sometimes try to
*match* the expression -- that's why some tests are in matcher.cpp
rather than filterparser.cpp.

We *do* examine the parse tree in Rust, though, so there the tests are
in their logical places.