Cirrus CI: run Clippy (Rust linter)
It's not a given that warnings would be the same on every platform; we
already ran into this with `libc::mode_t` having different size between
Linux and FreeBSD. That's why we opt to run Clippy in *all* tasks,
rather than creating a single standalone task just for linting.

Had to switch Rustup from "minimal" profile to "default" to install
Clippy.