Drop LogLevel argument from ScopeMeasure c-tor
Only tests specified that argument; the rest of the code relied on the
default. Rust doesn't have default arguments, though, relying instead on
the builder pattern. It's far better to drop code than to add it, so I'm
removing the argument altogether.

Should have done it before porting ScopeMeasure to Rust, but oh well.