Improve docs on sharing Cargo cache with Docker
Previously, we instructed people to mount the whole of host's ~/.cargo
into the container. This works fine if rustup is already installed on
the host: in this case, the contianer uses host's proxies in
~/.cargo/bin to locate the correct tools (still inside the container).
But if the host doesn't have rustup installed, the mount effectively
wipes ~/.cargo inside the container, so `cargo` can no longer be found.

This commit changes the suggestion to mounting ~/.cargo/registry. If it
doesn't exist on the host, the user can safely create an empty
directory, and that won't break anything.

Technically, we could also share ~/.cargo/git, but Newsboat doesn't
fetch crates from Git right now, so I didn't bother.

Cf. #1064.