FilterParser: test that Error::Internal never happens
`Error::Internal` means "we have no error message for that". This is
a bad form, so I added a test to ensure that this never actually
happens.

I evaluated 2.5 million cases (PROPTEST_CASES=2560000 cargo test), and
the test never failed. This doesn't *prove* that this error can't
happen, but at least gives me some assurance that it can't be triggered
trivially.