Add Miniflux docker-compose test file
With the new Miniflux integration it can be helpful to have a test
environment. This docker-compose brings up Miniflux and a PostgreSQL
database with a preconfigured setup. After bringing up the docker
environment, navigate to http://localhost and use the credentials set in
the docker-compose.yml to login and add a sample RSS feed.