Recommend explicit HTTP auth setting for Miniflux
Newsboat incorrectly guesses the type of HTTP authentication that should
be used with Miniflux, which results in an erronous unauthenticated HTTP
request for every Miniflux API call (effectively doubling the amount of
requests for no reason). Explicitly setting the authentication method to
`basic` forces Newsboat to always use the correct type on the first try.