Make RssFeed "First Tag" method ignore title tags
Made the get first tag method on the RssFeed class ignore title tags as
if those types of tags are the first on the tag array, then using any
kind of 'first tag' property on the application would have the same
effect as a 'title' one, removing any kind of usefullness of the
previous one. This is a problem when reading feeds from Inoreader for
example where the title tags are inserted first.

Changed existing tests not only for the rssfeed class getfirsttag()
method, but also for the feedcontainer class on the sorting by
firsttag case where title tags were added to check if it behaves like
intended.

Updated documentation on the title tags section to warn the users that
those types of tags are ignored when any 'first tag' property is used on
the application.

Fixes #1128