Fix query feeds with colons in quoted regex
This bugfix is inspired by
https://github.com/newsboat/newsboat/issues/1167. While trying to
reproduce that issue, I created the following query feed:

    "query:only interest:title =~ \"vivid:\""

To my surprise, Newsboat then failed to start:

    Error while loading feed 'query:only interest:title =~ "vivid:"': `title =~ "vivid:' is not a valid filter expression

Note how the closing quote is missing at the end of the filter
expression.

The problem was in RssFeed::set_rssurl(), which splits the query line
into three components: "query:" prefix, the title, and the filter
expression. set_rssurl() used utils::tokenize_quoted() for that, with
colon as a delimiter.

If we just split the query on a colon, we get four tokens:

1. query
2. only interest
3. title =~ "vivid
4. "

But "_quoted" means that the method "understands" double quotes: their
contents are treated as a single token. That method also has an
interesting property of implicitly closing the quotes. All of that
combined means that the fourth token above is treated as an empty
string, *not* as a stand-alone double quote. Combining the third token
with an (empty) fourth token, we then get a filter expression where the
quotes are not balanced, leading to an error.

This was trivial to fix by replacing utils::tokenize_quoted() with
utils::tokenize(), which doesn't "understand" double quotes.