Fix TOCTOU in Controller::mark_all_read()
Since this method is not holding a lock on FeedContainer, the `pos`
might become invalid in between the check and its use. We avoid this by
rewriting methods such that after the check, we have a pointer to the
`RssFeed` and thus don't need to worry about anything becoming invalid.

Spotted by @dennisschagt:
https://github.com/newsboat/newsboat/pull/1166#discussion_r483483403