Make utils::run_command() test less flaky
This test calls run_command("sleep", "5") and checks that the function
returns in less than 1 second, which should indicate that it *doesn't*
wait for `sleep` to finish running.

PR #1172 changed the implementation of utils::run_command(), which
caused this function to sometimes take more than a second to execute,
failing the test:

    test/utils.cpp:634: FAILED:
      REQUIRE( runtime.count() < 1000 )
    with expansion:
      1087 (0x43f) < 1000 (0x3e8)

I couldn't reproduce this locally even though I ran the tests in a loop
for half an hour. This makes me think that this is a peculiarity of the
CI environment, not an actual bug in utils::run_command().

Increasing the wait time to 5 seconds didn't help: the CI sometimes took
19 (nineteen!) seconds to continue executing the tests. So I increased
both the `sleep` argument and the wait time to 1 minute, and that seems
to have fixed it.

I fully realize this is still playing with probabilities; a day will
come when 60 seconds won't be enough either. I'll probably just remove
this test then, because it's not worth the effort.