Fix segfault in podboat --help
9c720413519aa29b1bb1ef3be9fff5e1696bba55 introduced a bug: `podboat
--help` segfaults after printing out the help message.

That commit split out some initialization code from
`PbController::run()` into a new method, `initialize()`. That new method
didn't *always* initialize the object; if --help or an unknown option
was passed into Newsboat, the method would print out the usage message
and return EXIT_SUCCESS.

However, this same EXIT_SUCCESS value was used to indicate that the
initialization was successful. As a result, Podboat proceeded to use
unitialized PbController, and crashed.

To fix this, I changed `PbController::initialize()` to call `exit(3)`
directly when it wanted to quit Podboat.