Docs: move dangling paragraphs into relevant chapters
After we turned tables of settings into lists, some chapters got
visually split in two: there's a part of the chapter, then a huge,
multi-screen list, then another part of the chapter. This commit
rearranges the text such that it always precedes tables, hopefully
making the documentation easier to read.