Use correct order of fields when specifying OP_INT_RESIZE mapping
"RESIZE" is the so-called event as returned from stfl (in the same way
as regular key-presses are registered).
So, "RESIZE" should be put in the `default_key` field.

The code which handled OP_INT_RESIZE was never called (before this fix).
Removing it in favor of a more generic approach when implementing the
idea from https://github.com/newsboat/newsboat/issues/390.