Continue execution, even when loading urls file fails
This fixes a regression introduced by
https://github.com/newsboat/newsboat/pull/1245.

Before the regression, the `urls` file could be omitted when importing
an OPML file or using an online service (e.g. miniflux). This commit
restores that behaviour. It still prints the error message to help users
diagnose issues in case of missing urls or if there are still 0 urls
configured.