Snap update: do not touch "edge" channel
Right now, we push releases to all channels simultaneously, and push Git
builds to "edge" only. When we rebuild an older version (to pick up new
libraries), it makes sense to push it everywhere *but* "edge", because
that channel already contains a newer version that's going to be updated
soon enough anyway.