Split QueueLoader::reload() into multiple methods
Signatures for these new methods are rather long, and so I ran into an
AStyle bug: the styles `attach-return-type` and `max-code-length` are
apparently in conflict. Given this code:

    RetType func(int x);
    //        ^
    //        |
    //        +-- let's say the length limit is here

AStyle applies the `max-code-length` style, and splits the line:

    RetType   // length
    func(     // limit
    int x);   // here

However, if we run AStyle again, its `attach-return-type` style comes
into effect and re-formats the code again:

    RetType func(
    int x);
    //        ^
    //        |
    //        +-- length limit

As you can see, the first line is over the limit again, and the next
AStyle invocation will break it apart, and the next one will glue it
back together, and so on.

The last time I ran into this in include/utils.h, I managed to avoid the
bug by replacing `RetType` with a short type synonym. In `QueueLoader`'s
case I can't do that because the method returns a private type which an
alias can't reference.

I reported the bug at https://sourceforge.net/p/astyle/bugs/545/, but
the fix will take a while (if it ever comes — there have been no
responses to issues for a year). So instead, I bumped the
`max-code-length` limit *just enough* for this code to fit. I plan to do
that again if we write more code that triggers this bug. I'm not happy
about this, since wide code is harder to put side-by-side with other
code, but it's the least intrusive change I can think of.