CI: add macOS Big Sur
We now use that release by default, since that's the current one, it's
less tested so far, and it's the one that more and more people will be
using at the time goes on.

We still keep one Catalina job to ensure compatibility, but I don't know
how long we'll keep doing that.