Move utils::is_valid_podcast_type() to cxx
We have to keep this one in the "bridged" module because
`nlohmann::json` doesn't want to get implicitly converted into
`rust::Str`, and *using* this utility starts being awkward: we have to
explicitly convert JSON into an `std::string` and store it in an
intermediate location for `rust::Str` to be happy. Bridge is much
simpler, and is universal to boot (write once, benefit everywhere).