Re-enable optimizations in CI builds
This effectively reverts f7f6d61cbdd5062e4bf3370e90f5f580ec12f1d6. The
second assumption in that commit was wrong: for newer compilers to show
the same warnings as the old ones, we should run them with the same
settings, but we don't. Newer compilers run with `-O3`, while the
default `make all` uses `-O2`. This led to
https://github.com/newsboat/newsboat/issues/1351, where a warning is
*only* shown at `-O2`, so we missed it.

By re-enabling optimization I'm slowing down the builds, but also ensure
that we actually check that Newsboat can be built in its default
configuration.