Work around -Wmaybe-uninitialized on GCC 9
This is a re-thinking of a454fff6a8b7456bd0d4c790c9911669daefa984 in
light of https://github.com/newsboat/newsboat/issues/1351. The problem
with that earlier commit is that `-Wno-unknown-warning-option` is only
supported by Clang. GCC treats it in a really bizarre way: it only
complains about unsupported warning when there are some other
diagnostics to report. Furthermore, if `-Werror` is in effect, there
seem to be no way to selectively demote the complaint back to
a warning — the only solution is `-Wno-error`.

So instead of playing with pragmas, I rewrote that code in a way that
doesn't trigger -Wmaybe-uninitialized. This is opportunistic, of
course — there is absolutely no guarantee that some future version of
some compiler, or even an existing compiler in a non-default build
configuration, won't emit the warning again. But we got CI checking the
default configuration now, so we should be able to catch the most
obvious cases.