Make cppcheck useable again
This adds a few suppression rules for files that we don't change
(3rd-party, auto-generated filter language parser, system-wide headers).

It also prevents those files from participating in cppcheck's
"configurations", which is essentially a set of macros that are defined
simultaneously. By cutting down the number of macros, I improved the
runtime of `make CPPCHECK_JOBS=9 cppcheck` from 41 minutes to 47
*seconds*.

Some of the test files can't be analysed because Catch's macros are too
opaque for cppcheck to handle. I didn't suppress the whole "tests"
directory because there are some seemingly useful warnings in there.