Add utils::locale_to_utf8()
It creates a nice symmetry with utils::utf8_to_locale(), and saves a bit
of typing in the calling code.

utils::convert_text() is no longer used outside of utils, but is kept
public because making it private is a pain: I'd have to copy its tests
to locale_to_utf8() and utf8_to_locale, which is just busywork.