Port utils::translit() to Rust
The "comparison test" which compares results of C++ and Rust
implementations now pass, and will be removed in the next commit.

This commit passed the CI, i.e. it works not just on my Linux machine,
but also on Ubuntu Linux with various compilers, two macOS versions, and
two FreeBSD versions.