Run C++ tests with time-based RNG seed
Turns out `--order=rand` does *not* provide random ordering on each
invocation[1] — we need `--rng-seed=time` for that. Performing tests in
randomized order is important because it helps us find and break
unintended dependencies between cases.

1. https://github.com/catchorg/Catch2/issues/2161

This change turned up a problem with tests for `utils::convert_text()`:
that function calls `utils::translit()`, which `abort()`s the whole
process if `iconv()` does not support the requested conversion. The
`abort()` is not a bug; there's nothing better we can do in that
situation. However, `abort()` is inconvenient for tests. I couldn't find
any better solution than deleting the offending test.

I ran this in a loop for 1545 times, and it never failed. This doesn't
prove much — to execute all 535 tests in all possible orders, I'll need
1e1221 ages of the Universe — but it's better than the previous record
of 53 runs :)