And a subsection re snap
As per an IRC discussion, just highlighting that, regardless of the Newsboat version (to date), references to `/usr/bin` will throw error code 127 as command not found, and references to `/home` will throw 126 as permission denied. But `xdg-open` is respected, and so offers an option for invoking the default browser without making further configuration changes.