Bump lexical-core to 0.7.5
This fixes the compile error that made 0.7.4 fail to build with recent
Nightly compilers (in turn failing our PRs #1459 and #1461). The change
in Nightly was reverted, giving lexical-core time to get fixed
(https://github.com/Alexhuszagh/rust-lexical/issues/55). Now that that's
done, we can expect the change to be re-introduced into Nightly, and so
we have to update to lexical-core 0.7.5 lest our build breaks again.