Fixed Italian translation feed list title format.
- The feed list title format wasn't showing properly with Italian localization.
In 8880ba431012964d62c8d798191995e53a58a99f I introduced a bug in the format
string used by the title removing a "?" from a conditional expression used
to change the string when a filter is used.
I've fixed it and it's showing the correct format with tags and filter as well
as the normal view.
- I removed an s from the english plural not needed in Italian
for that sentence.