CI: check for dbg! macro
This macro prints even in the release mode, so we should avoid it.

For now, the check is not very constrained: it examines all files, not
just the Rust code, and the simplistic regex can be triggered even by
a comment. OTOH I'm pretty sure this check *can't* miss a legitimate
dbg! call, so I prefer not to make it any more complicated at this time.