Bump cxx to 1.0
The only problem was a new restriction (which should be lifted in the
future): cxx only lets us share types that are defined in the current
crate. As a result, our libnewsboat-ffi crate can't share types from the
libnewsboat crate. To work around that, we wrap the type into
a single-field struct. This makes the bindings a bit more wordy, since
we have to access the field, but it's not too bad.

Fixes #1542.