Fix documentation error
Almost forgot this.

I went with the full api url, inconsistent with other integrations,
because a) FreshRSS has two apis with equal weight and so specifying
which is being used is more important, even if the fever api is not
implemented here, b) potential compatability with other json-variant
greader apis, and c) FreshRSS can be (mis)configured to have a different
url structure that might e.g. include a /p/ as part of that url