Do not use backtick in default values for settings
Text in between backticks is interpreted as commands, whose output
replaces the backticks. This trips people up when they copy-paste the
default value from the docs and discover that part of the value doesn't
show in the program. We avoid this by replacing backticks with single
quotes.

Rationale for single quotes (from
https://github.com/newsboat/newsboat/issues/1520#issuecomment-812909101):

> Let's use single quotes. Double quotes would be my first choice, but
> filters often end with a double quote, and the resulting title would
> look a bit weird. Single quotes avoid this problem.

There was an alternative idea to use colons and spaces, but I rejected
it:

> > Meanwhile, the same user on IRC suggests removing quotes entirely,
> > and using a format like this:
> >
> > > Maybe not use them, like: feedlist-title-format   " Feeds: %u unread, %t total.%?F?Filter: %F&?%?T? Tag: %T&?"
> > > it gives me this (for tag):  Feeds: 0 unread, 9 total. Tag: OpenBSD
>
> I think it's not a very good idea, since it makes entirely unclear
> there a filter/tag ends.

Fixes #1520.