Do not run `logger` doctest
That doctest is useless because it only runs some commands without
checking the output. Essentially, it only checks that the code compiles;
now, this behaviour is documented because the test is marked as
"no_run". We have a suite of integration tests to check that the logger
does produce output.