Speed up Rust tests by turning doctests into unit tests (#1574)
* Turn strwidth_stfl doctests into unit tests

I also updated C++ tests to match.

* Turn `is_http_url` doctests into unit tests

I also updated C++ tests to match.

* Turn `absolute_url` doctests into unit tests

* Turn `censor_url` doctests into unit tests

* Turn `quote_for_stfl` doctests into unit tests

* Turn `get_basename` doctests into unit tests

* Turn `substr_with_width` doctests into unit tests

* Turn `substr_with_width_stfl` doctests into unit tests

* Turn `run_interactively` doctests into unit tests

* Turn `run_non_interactively` doctests into unit tests

* Turn `gentabs` doctests into unit tests

* Do not run `fmtstrformatter` doctest

That test is not important because the module already has a thorough
test suite. It's not worth linking yet another binary just to run a few
more assertions.

* Do not run `logger` doctest

That doctest is useless because it only runs some commands without
checking the output. Essentially, it only checks that the code compiles;
now, this behaviour is documented because the test is marked as
"no_run". We have a suite of integration tests to check that the logger
does produce output.

* Docs: ask people to avoid Rust doctests

* Drop unneeded `return`