convert_text: guarantee that the output is in `tocode`
`convert_text()` used to become a no-op when the source and destination
encodings match. At a glance, this made sense, because, well, why not
skip unnecessary work?

It now transpires that the work *is* necessary, since nobody checks that
the input is encoded to `fromcode`. As a result, short-circuiting
`convert_text()` can return data that's not in `tocode`.

This isn't theoretical. `convert_text()` is not used directly; it's
a helper for `locale_to_utf8()` and `utf8_to_locale()`. As their names
imply, either source or destination encoding passed to `convert_text()`
would be UTF-8. The other encoding is detected from the locale and will
often be UTF-8 as well. As a result, in UTF-8 locales Newsboat will
accept *any* data and pretend that it's valid UTF-8.

The question is how to handle invalid input. `convert_text()` could
return an error, or it could paper over the issue by replacing invalid
characters with some replacement. Historically, Newsboat was doing the
latter, so that's what I went with, but I expect to revisit this
decision in the future.