Replace <curses.h> with <ncurses.h>
AFAIK Newsboat requires Ncurses specifically, so let's reflect that in
headers. With Ncurses, the headers are equivalent, so there's no change
to the behaviour or the code.

This commit is inspired by
https://github.com/newsboat/newsboat/pull/1580. The errors in that PR
were caused by compiler's inability to find ncurses at all, and
replacing headers won't fix that. But the PR made me look into this,
hence the mention.