Address part of liw's feedback on the README
Lars Wirzenius offers to review FOSS READMEs[1], so I submitted ours.
This commit addresses most of the feedback.

> A slight concern is that it doesn't explain, or link to an
> explanation, what "an RSS/Atom feed reader" is or what it's for.
> Perhaps adding words to the effect that it's good for following blogs
> and news sites? And link to
> https://en.wikipedia.org/wiki/News_aggregator for more information?

Done, although I expect my explanation of a "feed reader" can be
improved.

> A carefully arranged screenshot of a terminal with the software
> running would add some punch and give an indication of what the
> software is like.

Done. I also took an opportunity to replace the screenshot on our
website, as it was getting dated.

> I find the logo credit and licensing at the top to be distracting.
> Couuld they be moved to the end, with the licensing information of the
> software itself?

Agree, done. All the legalese is now at the end of the README.

> I like that […] there are […] pointers at […] where to get support,
> although I'm worried that support is needed. The word support could
> perhaps be tweaked: help, maybe? Support sounds so formal.

I'm in two minds about this. "Support" might sound formal, but "Help"
won't look as well as a heading IMHO.

1. https://liw.fi/readme-review/