Add a class that holds a UTF-8 string
...except the encoding is not actually checked.

The purpose of this is to mark, at the type level, which strings are in
UTF-8 and which aren't. Those that are using the new `Utf8String` type
are, and that's what we want to see in the majority of the cases. The
strings that are using `std::string` are basically vector<char>, and
should be treated on a case-by-case basis; that's what we want to move
away from.

Over the next however many commits, I'm going to gradually convert the
codebase to this new type. In the end, we'll see `Utf8String` everywhere
except for the few functions that deal with input and output. We'll
convert locale charset to `Utf8String` at the earliest opportunity, and
hold onto it until the very last moment before converting it back into
the locale charset.

The constructor doesn't validate input UTF-8 because that'd be
counter-productive right now. We *know* that we mishandle encodings
right now. The correct fix is to convert them on the boundary, where the
I/O is done. Since I'm going to introduce `Utf8String` from the inside
out, validation will only get in my way. Once `Utf8String` is used
universally, I'll introduce validation to guard against programming
errors (i.e. cases where we thought that the input is in UTF-8, but it
turned out to be invalid).