Replace MB_CUR_MAX with MB_LEN_MAX
TagSoupPullParser::decode_entity allocates an array to store
a multi-byte representation of a single character. The multi-byte
encoding to use depends on the locale, and the length of the array
depends on the encoding (because e.g. UTF-32 can represent anything with
four bytes, UTF-16 always takes up 2 bytes, and CP1251 always uses
a single byte, etc.) MB_CUR_MAX represents the maximum needed length of
the array for the current locale.

Since it depends on the locale, MB_CUR_MAX is actually a macro that
calls a function. As a result, the array is now a variable-length array,
i.e. its size is not known at runtime. This, in turn, makes it
impossible to protect with a stack protector (-fstack-protector). If the
user enabled `-Wstack-protector`, they'll get a warning about this.

Replacing MB_CUR_MAX with MB_LEN_MAX fixes the issue because the latter
is an actual constant. This replacement shouldn't cause any problems
because MB_LEN_MAX is always equal or greater than MB_CUR_MAX. It's also
not *that* much bigger — I think it only goes up to 6 (maximum number of
bytes that UTF-8 uses to represent a codepoint). We should have no
problem allocating 6 bytes on the stack.

Closes #1598.