Make NEWSBEUTER_PATH_SEP a string
...and rename it to NEWSBOAT_PATH_SEP

This makes it possible to use with Utf8String. We can't allow appending
a `char` to the `Utf8String` because a single `char` is not necessarily
a valid UTF-8 codepoint. We already allow constructing Utf8String from
a string literal, though, so converting NEWSBEUTER_PATH_SEP into
a string was the easiest solution.