CI: use -fstack-protector-all
This adds stack canaries to all functions. That bloats up the code, and
slows it down, but it ensures that when our tests smash the stack, we'll
know about it. (In practice, this measure is only as good as our test
coverage and variety, but still.)

ssp-buffer-size setting is no longer needed because -all adds the
protector regardless of how large the arrays in the function are.