{dir,file}browser: store filetype separately from the name
Original implementation converted filetype into an ASCII letter and
prepended it to the filename. This is clever and memory-efficient, but
once we migrate to UTF-8 strings, this will become a bit unwieldy
because accessing the first *byte* and the first *character* will become
different things. To forestall these problems, we replace the string
with a struct that holds the filetype and the name.

We also take the opportunity to make this code generic, and re-use it in
FileBrowserFormAction and DirBrowserFormAction (which previously held
copies of the same code).