CI: do not mention Rust version in jobs' names
GitHub changed the way required checks are added: I now have to *search*
through the latest checks, and select them from a dropdown of
suggestions. To add another one, I have to check again. This is terrible
compared to the earlier UI, where I could simply go down a list of
latest checks and select the ones I want to require.

By using "stable" and "MSRV" instead of the version number, this commit
ensures that the names don't change, sparing me an hour of my lifetime
managing this nonsense.