Shorten integration tests' names
Our GitHub Workflow failed in #1653 because the automatically generated
name for the object file was too long:

    Error: could not write output to /home/runner/work/newsboat/newsboat/target/debug/deps/configpaths_create_dirs_returns_false_if_xdg_config_dir_exists_but_data_dir_doesnt_and_couldnt_be_created-9170954e0544fc50.configpaths_create_dirs_returns_false_if_xdg_config_dir_exists_but_data_dir_doesnt_and_couldnt_be_created.2a36eb5b13484269-cgu.0.rcgu.o: File name too long

This is caused by us using pretty long names for integration tests.
These descriptive names come from the C++ tests, and I'd prefer to keep
them, since they aid understanding. So I made a compromise: the names
are just numbers now, but the test function itself still has a long
descriptive name. That way, `cargo test` prints out long, useful names,
which one can `git grep` for if one needs to look at the code.

An alternative approach would be to put all of module's tests into
a single file (e.g. configpaths.rs). But that will require us to be
careful about each tests' environment: chdir-ing into the initial
directory once the test finishes, clearing any environment variables,
resetting the locale and so on. In addition to that, we would also have
to introduce some global lock, to force Rust to run the test serially.
That's pretty inconvenient, so I opted for an easier, but hackier
solution.