Fix perf regression in mark-all-read with remote APIs
516392755439d84e3cf9531b44220b03eefa5759 (PR #1558) changed
mark-all-read to mark each item individually. That made mark-all-read
aware of the filters, but also made it slower. This was especially
noticeable with remote APIs, where mark-read latency is high.

We can't speed it up in the general case, but when there is no filter,
we can fall back onto the faster mark_all_read() method which marks the
entire feed in one go. I tested that this new code is indeed faster and
doesn't re-introduce #1364.

Reported by Oneiric on IRC.