Do not pre-fill key hints
Each dialog has a string at the bottom, showing some common keys. STFL
files give these strings some default values, but that is unnecessary
because the values are always overwritten when a dialog is created:

$ diff -ruN \
    <(find src -name '*formaction.cpp' | sort ) \
    <(find src -name '*formaction.cpp' -print0 | xargs -0 grep set_keymap_hints | sed 's#\.cpp.*$#.cpp#' | sort)
--- /proc/self/fd/11    2021-06-28 16:50:25.993400541 +0300
+++ /proc/self/fd/14    2021-06-28 16:50:25.993400541 +0300
@@ -1,12 +1,10 @@
 src/dialogsformaction.cpp
 src/dirbrowserformaction.cpp
-src/emptyformaction.cpp
 src/feedlistformaction.cpp
 src/filebrowserformaction.cpp
 src/formaction.cpp
 src/helpformaction.cpp
 src/itemlistformaction.cpp
 src/itemviewformaction.cpp
-src/listformaction.cpp
 src/selectformaction.cpp
 src/urlviewformaction.cpp

Out of 12 dialogs, only 2 don't overwrite hints `set_keymap_hints()`.
Those two dialogs are `EmptyFormAction`, which is never actually
displayed; and `ListFormAction`, which is a base class for other
dialogs. It's fine that those two don't overwrite the hints string.

An upcoming commit will introduce some STFL markup into hints. If we
kept the defaults, we'd have to update them to use the markup too,
otherwise they'd get misleading. Since that'd be busywork, we opt to
remove the defaults instead.