Add separate color style for dialog title
Up until now, the title bar and the hints bar shared a single style,
called `info`. Soon, the hints bar will get additional styles for keys
and their descriptions. So a question arose: should the two bars share
the same "base" style, or should they be split?

In favour of sharing the style are simplicity and backwards
compatibility: users who already specify `info` can just carry on, and
the colorscheme authors don't have to think how `swap-title-and-hints`
setting will change the look of the program.

In favour of separate styles is the fact that the hints bar will have
five different elements to style: keys, key separator (comma),
key-description separator (colon), description, and the overall
background. On 16-color terminals, this can be a challenge to style,
even when attributes are used. If hints used the same "background" style
as the title, it could make it harder to come up with a nice
colorscheme.

That's why I decided to split the styles. To preserve compatibility, the
new `title` style inherits from `info` when not set explicitly.