Move hint preparation to KeyMap
FormAction and PbView both contain the same code that prepared the text
for the hints bar. This commit moves that code into KeyMap.

It's not ideal; it'd be better to upgrade Podboat's code to use
FormAction, and just remove the code from PbView. But that's a much
bigger refactoring, so I opted for an easier option for now.

I also took an opportunity to replace `char*` with `std::string` in the
definition of `KeyMapHintEntry`.