utils::retrieve_url: Improve error reporting of curl_easy_perform
The current implementation did disregard the return value of
curl_easy_perform and simply returned the value of "buf" in all cases.
However, the value of buf is not really well defined in the case of
transmission errors.

This commit there checks the returned CURLcode for errors, and further
produces a meaningful error message to the log to aid debugging.
The latter comes with a caveat so: To produce the errormessage, we have
to clobber the value of CURLOPT_ERRORBUFFER, even for cached handles.