Cirrus CI: use `task:` key everywhere
Cirrus allows us to use custom keys like `32bit_ubuntu_task`, from which
Cirrus derives the name of the job (to be displayed in the UI). However,
we also provide `name` key, and now Cirrus complains that we've given
it two names to choose from. This commit replaces all our custom keys
with `task`, so Cirrus can stop complaining and just use what we specify
in `name`.