Do not use PRI* in calls to gettext
In C, it's ergonomic to write code like this:

    printf(gettext("foo" PRId32 "bar"), 42);

gettext (both the function and the tool) have support for this idiom:
the message is extracted as "foo<PRId32>bar", and the .mo files seem to
store it in annotated form as well.

Rust has no PRI* constants, and string literals are concatenated via the
`concat` macro. As a result, `xtr` (Rust's replacement for `xgettext`)
doesn't support this pattern at all.

To reconcile those differences, this commit replaces all uses of PRI*
inside gettext with code like this:

    prinln!(gettext("foo%sbar"), &42.to_string())

This is done even on the C++ side, to avoid mistakes when that code is
ported to Rust.

(There are no problem using PRI* *separately* from gettext, so we keep
doing it both in C++ and Rust.)