Upon quit from itemview current position should be last item
Unfortunately, I also realized that we cannot simply advance to the next
item as the concept of "next" depends on the direction user is
traversing articles.  In particular next is the wrong direction if user
is traversing articles in reverse.  Jump to a numered article or search
probaly implies no direction.