Do not examine current feed when jumping to next unread
This bug was present ever since this feature was implemented in
c26be624fdd9f48b26587b76fd001a1c4a2cc402. When the `visible_feeds` array
is empty, `list.get_position()` would return 0, which is not a valid
array index.

Other "jumping" methods are not affected by this:
- `jump_to_previous_unread_feed()` only uses the 0 index when the size
  of the array is 1;
- `jump_to_next_feed()` explicitly checks that the index is within an
  array;
- `jump_to_previous_feed()` explicitly checks that the index is above
  0 before decrementing it;
- similar methods in `ItemListFormAction` already implement the same
  measures as we do here.

Fixes #1734.