FeedList: do not call unread_item_count() twice
This method scans the entire feed and can get expensive.

I considered unifying the loops in FeedList to avoid code duplication,
but that complicates conditions and leaves 0- and 1-item feeds as
special cases. I'm not convinced the unified version is any simpler or
more maintainable, so I left the loops as-is for now.