Docs: mention that Inoreader API omits enclosures
This is inspired by https://github.com/newsboat/newsboat/issues/252#issuecomment-903343014.

The `warning-caption` thing adds a large emoji next to the warning.
Without this, Asciidoctor would use the text "WARNING" instead, which
doesn't look as eye-catching as an emoji. The recipe for this is taken
from [1] and [2]. The icon used is U+26A0 WARNING SIGN.

1. https://docs.asciidoctor.org/asciidoc/latest/blocks/admonitions/#using-emoji-for-admonition-icons
2. https://github.com/asciidoctor/asciidoctor.org/issues/571