Rephrase passwordeval/-file descriptions
Now that the settings are sorted alphabetically, the previous version
conveyed the impression that the `passwordfile` is more secure than the
`passwordeval`, when in fact, it is not. The beginnings of both
descriptions were swapped, but the "secure" part was dropped from the
`passwordfile`. Storing password in cleartext is not particularly safe.
To emphasize that a bit more, the "plaintext" was added, too.